fix(ci): prune go build cache before the build

+++ scripts/prune-go-cache.sh
@@ -4,11 +4,13 @@ set -o errexit
 set -o pipefail
 
 
+DRY_RUN="$1"
+
 GOCACHE=$(go env GOCACHE)
 SIZE=`du -sxm ${GOCACHE} | awk '{print $1}'`
 echo "GOCACHE size: ${SIZE}MB"
 
-if [ $SIZE -gt 3500 ]; then
+if [ "${DRY_RUN}" != "1" ] && [ "${TRAVIS_BRANCH}" == "master" ] && [ $SIZE -gt 3500 ]; then
   echo "Pruning GOCACHE at ${GOCACHE}"
   find "${GOCACHE}" -type f -delete
 fi
